Struct arithmetic::DenseMultilinearExtension

pub struct DenseMultilinearExtension<F>
where
    F: Field,
{
    pub evaluations: Vec<F>,
    pub num_vars: usize,
}

Stores a multilinear polynomial in dense evaluation form.

Fields
evaluations: Vec<F>

The evaluation over {0,1}^num_vars

num_vars: usize

Number of variables

Implementations

impl<F> DenseMultilinearExtension<F>
where F: Field,

pub fn from_evaluations_slice( num_vars: usize, evaluations: &[F], ) -> DenseMultilinearExtension<F>

Construct a new polynomial from a list of evaluations where the index represents a point in {0,1}^num_vars in little endian form. For example, 0b1011 represents P(1,1,0,1)

pub fn from_evaluations_vec( num_vars: usize, evaluations: Vec<F>, ) -> DenseMultilinearExtension<F>

Construct a new polynomial from a list of evaluations where the index represents a point in {0,1}^num_vars in little endian form. For example, 0b1011 represents P(1,1,0,1)

pub fn relabel_in_place(&mut self, a: usize, b: usize, k: usize)

Relabel the point in place by switching k scalars from position a to position b, and from position b to position a in vector.

This function turns P(x_1,...,x_a,...,x_{a+k - 1},...,x_b,...,x_{b+k - 1},...,x_n) to P(x_1,...,x_b,...,x_{b+k - 1},...,x_a,...,x_{a+k - 1},...,x_n)

pub fn iter(&self) -> Iter<'_, F>

Returns an iterator that iterates over the evaluations over {0,1}^num_vars

pub fn iter_mut(&mut self) -> IterMut<'_, F>

Returns a mutable iterator that iterates over the evaluations over {0,1}^num_vars