Struct arithmetic::VirtualPolynomial

pub struct VirtualPolynomial<F: PrimeField> {
    pub aux_info: VPAuxInfo<F>,
    pub products: Vec<(F, Vec<usize>)>,
    pub flattened_ml_extensions: Vec<Arc<DenseMultilinearExtension<F>>>,
    /* private fields */
}

A virtual polynomial is a sum of products of multilinear polynomials; where the multilinear polynomials are stored via their multilinear extensions: (coefficient, DenseMultilinearExtension)

Number of products n = polynomial.products.len(),

Number of multiplicands of ith product m_i = polynomial.products[i].1.len(),

Coefficient of ith product c_i = polynomial.products[i].0

The resulting polynomial is

$$ \sum_{i=0}^{n} c_i \cdot \prod_{j=0}^{m_i} P_{ij} $$

Example: f = c0 * f0 * f1 * f2 + c1 * f3 * f4 where f0 … f4 are multilinear polynomials

flattened_ml_extensions stores the multilinear extension representation of f0, f1, f2, f3 and f4

products is [ (c0, [0, 1, 2]), (c1, [3, 4]) ]

raw_pointers_lookup_table maps fi to i

Fields
aux_info: VPAuxInfo<F>

Aux information about the multilinear polynomial

products: Vec<(F, Vec<usize>)>

list of reference to products (as usize) of multilinear extension

flattened_ml_extensions: Vec<Arc<DenseMultilinearExtension<F>>>

Stores multilinear extensions in which product multiplicand can refer to.

Implementations

impl<F: PrimeField> VirtualPolynomial<F>

pub fn new(num_variables: usize) -> Self

Creates an empty virtual polynomial with num_variables.

pub fn new_from_mle( mle: &Arc<DenseMultilinearExtension<F>>, coefficient: F, ) -> Self

Creates an new virtual polynomial from a MLE and its coefficient.

pub fn add_mle_list( &mut self, mle_list: impl IntoIterator<Item = Arc<DenseMultilinearExtension<F>>>, coefficient: F, ) -> Result<(), ArithErrors>

Add a product of list of multilinear extensions to self Returns an error if the list is empty, or the MLE has a different num_vars from self.

The MLEs will be multiplied together, and then multiplied by the scalar coefficient.

pub fn mul_by_mle( &mut self, mle: Arc<DenseMultilinearExtension<F>>, coefficient: F, ) -> Result<(), ArithErrors>

Multiple the current VirtualPolynomial by an MLE:

add the MLE to the MLE list;

multiple each product by MLE and its coefficient. Returns an error if the MLE has a different num_vars from self.

pub fn evaluate(&self, point: &[F]) -> Result<F, ArithErrors>

Evaluate the virtual polynomial at point point. Returns an error is point.len() does not match num_variables.

pub fn rand<R: RngCore>( nv: usize, num_multiplicands_range: (usize, usize), num_products: usize, rng: &mut R, ) -> Result<(Self, F), ArithErrors>

Sample a random virtual polynomial, return the polynomial and its sum.

pub fn rand_zero<R: RngCore>( nv: usize, num_multiplicands_range: (usize, usize), num_products: usize, rng: &mut R, ) -> Result<Self, ArithErrors>

Sample a random virtual polynomial that evaluates to zero everywhere over the boolean hypercube.

pub fn build_f_hat(&self, r: &[F]) -> Result<Self, ArithErrors>

pub fn print_evals(&self)

Print out the evaluation map for testing. Panic if the num_vars > 5.